+++ arch/arm/mach-tegra/tegra12_emc.c
@@ -330,25 +330,25 @@ static void __iomem *clk_base = IO_ADDRESS(TEGRA_CLK_RESET_BASE);
 
 static inline void emc_writel(u32 val, unsigned long addr)
 {
-	writel(val, (u32)emc_base + addr);
+	writel(val, emc_base + addr);
 }
 
 static inline u32 emc_readl(unsigned long addr)
 {
-	return readl((u32)emc_base + addr);
+	return readl(emc_base + addr);
 }
 static inline void mc_writel(u32 val, unsigned long addr)
 {
-	writel(val, (u32)mc_base + addr);
+	writel(val, mc_base + addr);
 }
 static inline u32 mc_readl(unsigned long addr)
 {
-	return readl((u32)mc_base + addr);
+	return readl(mc_base + addr);
 }
 static inline void ccfifo_writel(u32 val, unsigned long addr)
 {
-	writel(val, (u32)emc_base + EMC_CCFIFO_DATA);
-	writel(addr, (u32)emc_base + EMC_CCFIFO_ADDR);
+	writel(val, emc_base + EMC_CCFIFO_DATA);
+	writel(addr, emc_base + EMC_CCFIFO_ADDR);
 }
 
 static int last_round_idx;
